Move utility function to its own module
